style: make journal code internal and in separate package
