+++ bugsnag-android-core/src/main/java/com/bugsnag/android/App.kt
@@ -1,8 +1,8 @@
 package com.bugsnag.android
 
 import com.bugsnag.android.internal.ImmutableConfig
-import com.bugsnag.android.internal.JournalKeys
-import com.bugsnag.android.internal.Journalable
+import com.bugsnag.android.internal.journal.JournalKeys
+import com.bugsnag.android.internal.journal.Journalable
 import java.io.IOException
 
 /**
