+++ bugsnag-android-core/src/main/java/com/bugsnag/android/Notifier.kt
@@ -1,7 +1,7 @@
 package com.bugsnag.android
 
-import com.bugsnag.android.internal.JournalKeys
-import com.bugsnag.android.internal.Journalable
+import com.bugsnag.android.internal.journal.JournalKeys
+import com.bugsnag.android.internal.journal.Journalable
 import java.io.IOException
 
 /**
