+++ bugsnag-android-core/src/test/java/com/bugsnag/android/JournalTest.kt
@@ -1,11 +1,12 @@
 package com.bugsnag.android
 
-import com.bugsnag.android.internal.Journal
+import com.bugsnag.android.internal.journal.Journal
 import org.junit.Assert
 import org.junit.Test
 import java.io.ByteArrayOutputStream
 
-class JournalTest {
+internal class JournalTest {
+
     @Test
     fun testEmptyEverything() {
         val journal = Journal(standardType, standardVersion)
