+++ bugsnag-android-core/src/test/java/com/bugsnag/android/JournaledDocumentTest.kt
@@ -1,8 +1,8 @@
 package com.bugsnag.android
 
-import com.bugsnag.android.internal.Journal
-import com.bugsnag.android.internal.JournaledDocument
-import com.bugsnag.android.internal.JsonHelper
+import com.bugsnag.android.internal.journal.Journal
+import com.bugsnag.android.internal.journal.JournaledDocument
+import com.bugsnag.android.internal.journal.JsonHelper
 import org.junit.Assert
 import org.junit.Rule
 import org.junit.Test
